+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New method `Redmine\Api\Attachment::fromHttpClient()` for creating the class.
 - New method `Redmine\Api\CustomField::fromHttpClient()` for creating the class.
+- New method `Redmine\Api\Group::fromHttpClient()` for creating the class.
 - New method `Redmine\Api\Issue::fromHttpClient()` for creating the class.
 - Add support for PHP 8.5
 - Add support for Redmine 6.1.
@@ -25,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Extending `Redmine\Api\Attachment` is deprecated and will be set to final in future, create a wrapper class instead.
 - `Redmine\Api\CustomField::__construct()` is deprecated and will be set to private in future, use `\Redmine\Api\CustomField::fromHttpClient()` instead.
 - Extending `Redmine\Api\CustomField` is deprecated and will be set to final in future, create a wrapper class instead.
+- `Redmine\Api\Group::__construct()` is deprecated and will be set to private in future, use `\Redmine\Api\Group::fromHttpClient()` instead.
+- Extending `Redmine\Api\Group` is deprecated and will be set to final in future, create a wrapper class instead.
 - `Redmine\Api\Issue::__construct()` is deprecated and will be set to private in future, use `\Redmine\Api\Issue::fromHttpClient()` instead.
 - Extending `Redmine\Api\Issue` is deprecated and will be set to final in future, create a wrapper class instead.